Redesigned the jfmi.database package classes.
DatabaseRecord: has changed from an interface to an abstract class. This is
so that the DatabaseRecord class can specify some static Strings representing
class-common SQL statements, but it to subclasses to initialize them.

RecordConverter: is a new interface. Classes which implement this interface
are stating that they will implement methods to:
	1) Convert a row in a ResultSet from a query on table X to an appropriate
	Java DatabaseRecord subclass instance.
	2) [TODO] Convert a DatabaseRecord subclass instance to an SQL statement
	which will insert a matching record into table X.

FileRecord, TagRecord, TaggingRecord: have all been updated to reflect the
changes to DatabaseRecord.

FileRecordConverter, TagRecordConverter, TaggingRecordConverter: have been
newly created to provide conversion for DatabaseRecord subclasses.