SQLiteDatabase.getAllTaggedFiles() is fixed
The problem was that it called FileRecord.getSelectAllSQL() to get an SQL
statement. It then created a Statement and executed the SQL. But,
since FileRecord inherited the static method getSelectAllSQL() from
DatabaseRecord, and did not hide it, DatabaseRecord.getSelectAllSQL() ended
up being called.
This resulted in the static String DatabaseRecord.selectAllSQL being returned,
instead of FileRecord.selectAllSQL string.
Problem was fixed in FileRecord (and in TagRecord, TaggingRecord) by providing
proper static method implementations to hide those provided by DatabaseRecord.