New methods for managing an edited TaggedFile
The methods assignAdded(), assignRemoved(), assignUpdated() have been added
to the EditedTaggedFile class. These methods assign a FileTagging to the
specified set if it is possible to do so, and if so the methods maintain
consistency among the three sets.

NOTE the methods are not implemented, just sketched out in pseudocode for
now