EditedTaggedFile uses FileTaggingComparator
EditedTaggedFile uses FileTaggingComparator for its sets now. The
FileTaggingComparator compares FileTagging instances based on the primary
keys of the underlying storage repository. This ensure that we don't get
duplicates, which the natural ordering could result in (e.g. objects with
the same primary keys, but with other fields that are different, are
considered the same object by the FileTaggingComparator, but different by
the natural ordering).